Rename namespace and fix reference
