+++ Projekt/NoNameBooking/ApplicationClassLibrary/PractitionerViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace ViewModelClassLibrary
+﻿namespace ApplicationClassLibrary
 {
     public class PractitionerViewModel
     {
