+++ Projekt/NoNameBooking/ApplicationClassLibrary/UserViewModel.cs
@@ -1,9 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Text;
 using ModelClassLibrary;
 
-namespace ViewModelClassLibrary
+namespace ApplicationClassLibrary
 {
     public class UserViewModel
     {
