+++ Projekt/NoNameBooking/BookNyAftale/BookNyAftale.csproj
@@ -116,13 +116,9 @@
     <None Include="App.config" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\ModelClassLibrary\ModelClassLibrary.csproj">
-      <Project>{BB1603B7-1593-4B0B-BE66-EC29CF1635DD}</Project>
-      <Name>ModelClassLibrary</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\ViewModelClassLibrary\ViewModelClassLibrary.csproj">
-      <Project>{0EF19BAF-D374-4C52-8DF6-093E45A2F2B7}</Project>
-      <Name>ViewModelClassLibrary</Name>
+    <ProjectReference Include="..\ApplicationClassLibrary\ApplicationClassLibrary.csproj">
+      <Project>{77ad1380-e0f9-44f1-a870-d8d47ca8982c}</Project>
+      <Name>ApplicationClassLibrary</Name>
     </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
