+++ Projekt/NoNameBooking/BookNyAftale/CreateAppointment.xaml.cs
@@ -12,7 +12,7 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
-using ViewModelClassLibrary;
+using ApplicationClassLibrary;
 
 
 namespace BookNyAftale
